+++ tests/core/test_thunks.py
@@ -2,12 +2,11 @@
 
 import pytest
 
-from aioway._core import TensorInput, Thunk
+from aioway._core import TensorInput, Thunk, TensorThunk
 from aioway.modes import (
     Aten,
     AtenThunk,
     HistTensorGraph,
-    ModeThunk,
     NnFwdThunk,
     NnInitThunk,
     TorchDispThunk,
@@ -27,7 +26,7 @@ def _fn_cls():
 
 def _input_cls():
     yield HistTensorGraph
-    yield ModeThunk
+    yield TensorThunk
     yield AtenThunk
     yield NnFwdThunk
     yield AtenThunk
